Remove documentation install logic
The logic to install the documentation files never worked, and
never could have worked, as it's based on a misunderstanding of
CMake. The glob that looks for documentation files will be run
at build _generation_ time, when the doc files have not been
built yet, so it will never find any.

Installing files based on a glob is what the `FILES MATCHING`
argument to `install()` is for. (But we can't use that, either,
since generating doc files is optional and won't necessarily be
done before install. It's fine to not install the docs, they
can always be browsed from the build directory, or packaged
right from there.)