CMake: REQUIRE OpenMP, use (or create) targets
OpenMP is now REQUIRED as the build will fail if it's not available.
This PR updates the `find_package`, and creates the necessary targets
for older CMake versions (3.9+ will do it automatically).

The targets are then made PUBLIC dependencies of the openshot
target, which causes the OpenMP configs to automatically propagate.
The redundant OpenMP detection is removed from test/CMakeLists.txt.

The detection code comes verbatim from this post, which also credits
Modern CMake: https://iscinumpy.gitlab.io/post/omp-on-high-sierra/