Keyframe tests: Add test about large segment, including performance
This new test makes sure that a large segment is handled correctly and
also with reasonable performance.  The timeout of 10ms is still
relatively slow (on a mid-class laptop it takes 0.001ms currently) but
the test shouldn't fail when e.g. the build machine is under (mild to
heavy) load.