Modernize FindUnitTest++.cmake, add pkg-config
The same old variables will be respected for setting the path, but the
PREFERRED method is defining `-DUnitTest++_ROOT=/path/to` on the
CMake command line. _ROOT variables are handled with extra intelligence
by CMake.

The find module will also attempt to locate the pkg-config file
UnitTest++.pc, and if found will import its data.