Fixed bug in ObjectDetection effect
The bug made Openshot crash if two or more ObjectDetection effects were added to the any clip on the timeline.
The bug was solved by removing the static int kf_count variable from KalmanTracker. This variable was being incremented even if the ObjectDetection effect was removed from the clip, that caused the wrong Tracked Objects indexes to be added to the trackedObjects map.