CMake: Add separate unit test targets
Each unit test file ClassName.cpp, which results in the creation
of a test target openshot-ClassName-test, will now also be labeled
with the CTest label ClassName, and a target ClassName_coverage
will be generated that runs only the tests in that file.

This is especially useful when developing tests for a class, as
the tests in the class being worked on can be re-run without
having to wait a minute or more while the other tests run.