FFmpegWriter: Avoid some ugly #ifdef+fallthrough
There was some apparently-intentional use of switch case fallthrough,
done in concert with #ifdef wrapping to restrict code to certain
FFmpeg versions.

Fallthrough is too magical at the best of times. (In fact it'll
cause compiler warnings, with -Wall -Wextra. Unless it's in the
form of simple stacked case:s, which are considered OK).

#ifdef inside switch (effectively creating mutable switch
statements) is pretty bad, too. Combining the two is a recipe
for disaster.

So, those sections of code are replaced with equivalent
simple if()s, wrapped in the same #ifdefs. One switch still
uses an #ifdef with fallthrough, but only in the simple
stacked-case form:

    switch (condition) {
#ifdef
    case MAYBETHING :
#endif
    case THING2 :
    CASE THING3 :
    case THING4 :
        // stuff here...
        break;

...Which I think is clear enough at face value to be OK.