Avoid a bug in ImageMagick or cassert headers
At least on Debian bullseye as at today, the ImageMagick headers
include <cassert> with an ImageMagick namespace current. <cassert>
and the headers it then pulls in also do not set the namespace.
Those headers then declare functions such as __assert_fail, which
end up in the ImageMagick namespace. This causes failures in other
parts of the build, such as the ZeroMQ code. This seems to be a
bug in either the ImageMagick headers or the <cassert> header, but
can be easily worked around by including <cassert> before any
ImageMagick header. This change does so.