QtImageReader: Create from QImage
To add flexibility, a new constructor is added that accepts a
QImage (by const reference — NOTE: NOT a std::shared_ptr),
and stores that image as its source instead of loadin from
disk.

This isn't directly useful to OpenShot (it can't be configured
via the reader's JSON interface), but came up in a discussion
on unit testing as a simpler alternative to the DummyReader.
It may also be useful to other C++ consumers of libopenshot.