Catch2: Header templates for tests
Since Catch2 v3.0+ uses a vasty different organization for headers
and targets, we have two templates for what the unit test code
will include. One or the other is configure by CMake as the file
openshot_catch.h in the build directory, and all of the unit test
sources can include that instead.