Fixes test file collisions with high concurrency builds Fix by Ninpo <ninpo@qap.la>
When tests are run in parallel it's possible for ImageWriter
and FFmpegWriter Gif tests to truncate/overwrite the other's
output1.gif. This change makes use of test file names unique
to the tests themselves.

Because cmake defaults to Makefiles and make does not -j$(nproc)
by default, this bug can't show up unless part of a build with parallel
make with tests included.

https://github.com/OpenShot/libopenshot/issues/933